+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Delete a dataset using the command line
 - Code to parse VCF files (SNVs) and create Variant objects
 - Save SNV variants from parsed VCF files
+- Update SNV variants for one or more VCF samples